api/requirements.txt
Flask==1.1.1
Flask_SQLAlchemy==2.4.1
flask-cors==3.0.7
flask_marshmallow==0.10.1
marshmallow-sqlalchemy==0.21.0
PyJWT==1.7.1
